pyproject.toml
[project]
name = "nanoCMS"
version = "0.1.0"
description = "nanoCMS - A lightweight CMS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cerberus>=1.3.7",
    "flask>=3.1.0",
    "flask-cors>=6.0.1",
    "flask-limiter>=3.13",
    "flask-wtf>=1.2.2",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.3",
]
